refactor: replace vendored op-alloy-network with inline Optimism type
Define a minimal `Optimism` network type directly in op-steel instead
of maintaining a patched fork of op-alloy-network. This eliminates the
`patches/op-alloy-network` directory and both `[patch.crates-io]`
entries while preserving full functionality.

The upstream fix (ethereum-optimism/optimism#19756) for the
`NetworkWallet<Optimism>` conflict has been merged but not yet released.
Tracked by #116.
