+++ Cargo.toml
@@ -36,7 +36,6 @@ alloy-sol-types = { version = "1.3" }
 # OP Steel
 alloy-op-evm = { version = "0.30" }
 op-alloy-consensus = { version = "0.24" }
-op-alloy-network = { version = "0.24" }
 op-alloy-rpc-types = { version = "0.24" }
 
 # Alloy host dependencies
@@ -60,10 +59,3 @@ test-log = "0.2.15"
 thiserror = { version = "2.0" }
 tokio = { version = "1.35" }
 url = { version = "2.5" }
-
-# Patch op-alloy-network to remove the conflicting `NetworkWallet<Optimism>` impl
-# that clashes with the blanket impl added in alloy-network 1.8.
-# See: https://github.com/ethereum-optimism/optimism/pull/19756
-# Remove this patch once a new op-alloy-network is published with the fix.
-[patch.crates-io]
-op-alloy-network = { path = "patches/op-alloy-network" }
