--- patches/op-alloy-network/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-edition = "2024"
-rust-version = "1.88"
-name = "op-alloy-network"
-version = "0.24.0"
-authors = ["Alloy Contributors"]
-description = "Optimism blockchain RPC behavior abstraction"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/ethereum-optimism/optimism"
-
-[features]
-serde = [
-    "op-alloy-consensus/serde",
-    "op-alloy-rpc-types/serde",
-    "alloy-consensus/serde",
-    "alloy-primitives/serde",
-    "alloy-rpc-types-eth/serde",
-]
-std = [
-    "op-alloy-consensus/std",
-    "op-alloy-rpc-types/std",
-    "alloy-consensus/std",
-    "alloy-primitives/std",
-    "alloy-rpc-types-eth/std",
-]
-
-[dependencies]
-alloy-consensus = { version = "1.6.3", default-features = false }
-alloy-network = { version = "1.6.3", default-features = false }
-alloy-primitives = { version = "1.5.6", default-features = false, features = ["map-foldhash"] }
-alloy-provider = { version = "1.6.3", default-features = false, features = ["reqwest", "debug-api"] }
-alloy-rpc-types-eth = { version = "1.6.3", default-features = false }
-op-alloy-consensus = { version = "0.24.0", default-features = false, features = ["alloy-compat"] }
-op-alloy-rpc-types = { version = "0.24.0", default-features = false }
